+++ projects/insert-order-service/src/lib/view-orders/1#view-order-services/view-orders-graphql-strings.service.ts
@@ -7,7 +7,7 @@ import { gql } from 'apollo-angular-boost';
 export class ViewOrdersGraphqlStringsService {
 
     public GET_MEDIUM_DATA_FOR_SPECIFIC_ROUTE = gql`
-    query getSpecificRoute_MEDIUM($accountid:Float, $timeStampid:Float, $routeid: Float){
+    query getSpecificRoute_MEDIUM($accountid:Int, $timeStampid:Int, $routeid: Int){
         nodeOrderDetailsMicroService(timeStampid:$timeStampid, routeid:$routeid, accountid:$accountid){
             edges{
                 node{
